organizando y separando logica del codigo
